synq: split lsp host into document store + hover/completion services
`LspHost` was a god module mixing document storage, analysis dispatch,
hover/definition shaping, and completion assembly. Split into thin
delegators over three new modules:

- `DocumentStore` (document_store.rs): owns the URI → Document map and
  its open/update/close/invalidate lifecycle. `Document` + its caches
  move here.
- `hover_service.rs`: free functions for `hover_info`, `definition_info`,
  `prepare_rename`, `symbol_identity_at`, and hover formatting — pure
  transforms over `DocumentAnalysisData`.
- `completion_service.rs`: `build_completion_items` assembles the entry
  list from a pre-computed `CompletionInfo` + dialect + catalog.

Public API is unchanged; all delegating methods on `LspHost` keep their
signatures. `ensure_analysis_for` now takes `&mut DocumentStore` instead
of `&mut HashMap<String, Document>`.
